final refactor (hopefully).
